feat: Configure display function
Make the exact output to the echo area configurable. The default display
preserves the existing behaviour (clean leading and trailing colons,
find relevant expression using treesitter).

This change also requires moving some of the logic from the
pre-processing phase to the display phase (namely, the treesitter part),
which also makes it easier to apply a couple more highlight groups to
the results.